+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 #### v1.16.0 `2017-??-??`
 - `Fixed`
     - Removed EOL PHP versions (i.e. `5.4` and `5.5`). Done by [@raphaelstolt](https://github.com/raphaelstolt). See [#211](https://github.com/jonathantorres/construct/issues/211).
+    - Added a guard to check if the project directory to be already exists. Done by [@raphaelstolt](https://github.com/raphaelstolt). See [#212](https://github.com/jonathantorres/construct/issues/212).
 
 #### v1.15.2 `2017-09-18`
 - `Fixed`